+++ packages/core/lib/core.dart
@@ -26,10 +26,16 @@ export 'src/responsive/responsive_layout.dart';
 export 'src/router/app_navigator_observer.dart';
 export 'src/router/app_routes.dart';
 
+// services
+export 'src/services/firebase_service.dart';
+export 'src/services/fcm_notification_service.dart';
+
 // storage
 export 'src/storage/secure_storage_service.dart';
 export 'src/storage/shared_preferences_storage.dart';
 export 'src/storage/storage_service.dart';
+export 'src/storage/app_database.dart';
+export 'src/storage/app_database_provider.dart';
 
 // theme
 export 'src/theme/app_colors.dart';